+++ src/test/converter2/issues/gh2020.js
@@ -1,0 +1,7 @@
+// @ts-check
+
+/**
+ * @typedef {Object} Options Desc
+ * @property {string} url - Desc2
+ * @property {string} [apiKey] - Desc3
+ */