+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+### Bug Fixes
+
+-   Fixed schema URL for TSDoc preventing the use of `typedoc/tsdoc.json` in TSDoc extends, #2016.
+-   Fixed missing JSDoc style `@typedef` comments for properties, #2020.
+
 ## v0.23.9 (2022-07-24)
 
 ### Bug Fixes