+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 -   Fixed schema URL for TSDoc preventing the use of `typedoc/tsdoc.json` in TSDoc extends, #2016.
 -   Fixed missing JSDoc style `@typedef` comments for properties, #2020.
 
+### Thanks!
+
+-   @nazarhussain
+
 ## v0.23.9 (2022-07-24)
 
 ### Bug Fixes